서버 DB에 있는 댓글과 답글 갯수를 사용해서 표시하고, 없는 경우에만 직접 client에서 조회
