+++ src/hooks/usePosts.ts
@@ -1,9 +1,9 @@
-import { DocumentData, limit, QueryDocumentSnapshot } from "firebase/firestore";
-import { query, collection, orderBy, where, getDocs, startAfter } from "firebase/firestore";
+import { limit, query, collection, orderBy, where, getDocs, startAfter } from "firebase/firestore";
 import { firestore } from "@/firebase";
 import { Post } from "@/types/Posts";
 import { useInfiniteQuery } from "@tanstack/react-query";
 import * as Sentry from '@sentry/react';
+import { mapDocToPost } from "@/utils/mapDocToPost";
 
 export const usePosts = (boardId: string, selectedAuthorId: string | null, limitCount: number) => {
     return useInfiniteQuery<Post[]>(
@@ -42,35 +42,6 @@ async function fetchPosts(boardId: string, selectedAuthorId: string | null, limi
     }
 
     const snapshot = await getDocs(q);
-    const postsData = await Promise.all(snapshot.docs.map((doc) => mapDocToPost(doc, boardId)));
+    const postsData = await Promise.all(snapshot.docs.map((doc) => mapDocToPost(doc)));
     return postsData;
-}
-
-async function mapDocToPost(docSnap: QueryDocumentSnapshot<DocumentData>, boardId: string): Promise<Post> {
-    const data = docSnap.data();
-    return {
-        id: docSnap.id,
-        boardId: data.boardId,
-        title: data.title,
-        content: data.content,
-        authorId: data.authorId,
-        authorName: data.authorName,
-        comments: await getCommentsCount(boardId, docSnap.id),
-        createdAt: data.createdAt?.toDate() || new Date(),
-        updatedAt: data.updatedAt?.toDate(),
-        weekDaysFromFirstDay: data.weekDaysFromFirstDay,
-    };
-}
-
-async function getCommentsCount(boardId: string, postId: string): Promise<number> {
-    const commentsSnapshot = await getDocs(collection(firestore, `boards/${boardId}/posts/${postId}/comments`));
-    const commentsCount = await Promise.all(
-        commentsSnapshot.docs.map(async (comment) => {
-            const repliesSnapshot = await getDocs(
-                collection(firestore, `boards/${boardId}/posts/${postId}/comments/${comment.id}/replies`),
-            );
-            return Number(comment.exists()) + repliesSnapshot.docs.length;
-        }),
-    );
-    return commentsCount.reduce((acc, curr) => acc + curr, 0);
 }
